Moved analysis class to seperate file and added custom exception
